Branch off Ricky to display some test cases
This branch is a version of Ricky branch for the purpose of test cases or showcasing for HTTP Get connection (requests and responses) and multi-threading. Threads were created manually in the client.py. Also, the proxy_server.py has a sleep function in proxyThreading() to allow the console to display multithreading among the threads.